refactor repl packing, use -impl (suffix) to the classes and remove I- (prefix) from interfaces
