--- .idea/shelf/Uncommitted_changes_before_Merge_at_04_12_2022_9_51_pm__Changes_.xml
@@ -1,4 +1,0 @@
-<changelist name="Uncommitted_changes_before_Merge_at_04_12_2022_9_51_pm_[Changes]" date="1670183481516" recycled="true" deleted="true">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Merge_at_04_12_2022_9_51_pm_[Changes]/shelved.patch" />
-  <option name="DESCRIPTION" value="Uncommitted changes before Merge at 04/12/2022 9:51 pm [Changes]" />
-</changelist>
