+++ app/src/main/java/io/github/mohamed/sallam/awb/db/entity/DetoxPeriod.java
@@ -8,11 +8,11 @@ import java.util.Date;
 
 import io.github.mohamed.sallam.awb.db.converter.TimestampConverter;
 import io.github.mohamed.sallam.awb.db.converter.UuidConverter;
-import io.github.mohamed.sallam.awb.repo.IAggregateRoot;
+import io.github.mohamed.sallam.awb.repo.AggregateRoot;
 
 @TypeConverters({UuidConverter.class, TimestampConverter.class})
 @Entity(tableName = "detox_periods_table")
-public class DetoxPeriod extends DetoxSettings implements IAggregateRoot {
+public class DetoxPeriod extends DetoxSettings implements AggregateRoot {
 	// Fields
 	public Timestamp endDate = new Timestamp(0);
