+++ app/src/main/java/io/github/mohamed/sallam/awb/repo/DeviceRepository.java
@@ -1,90 +1,55 @@
 package io.github.mohamed.sallam.awb.repo;
 
-import android.app.Application;
-
 import androidx.lifecycle.LiveData;
 
 import java.util.List;
 import java.util.UUID;
 
-import io.github.mohamed.sallam.awb.db.UserDatabase;
-import io.github.mohamed.sallam.awb.db.dao.DeviceDao;
-import io.github.mohamed.sallam.awb.db.dao.GroupDao;
 import io.github.mohamed.sallam.awb.db.entity.Device;
 import io.github.mohamed.sallam.awb.db.relationship.DeviceWithGroups;
 
 /**
- * {@inheritDoc}
+ * {@link IRepository}
  *
- * {@link IDeviceRepository}
+ * `IDeviceRepository` interface is responsible for handling and declaring
+ * device operations (insert - update - delete - generateUuid - getters). On class
+ * `DeviceRepository` we add the full logic and implementation to execute
+ * a specific operation by using DAOs methods.
  *
  * @author Abdalrhman Hemida
  * @author Mohamed Yehia
  */
-public class DeviceRepository implements IDeviceRepository {
-    private final DeviceDao deviceDao;
-    private final GroupDao groupDao;
-    private final LiveData<List<Device>> devices;
-    private final LiveData<List<DeviceWithGroups>> devicesWithGroups;
-    private final LiveData<Device> thisDevice;
+public interface DeviceRepository extends IRepository<Device> {
+    /**
+     * Updates device in database.
+     *
+     * @param device object of the blocking period.
+     */
+    void update(Device device);
 
     /**
-     * Instantiate an object from `DeviceDao` and `GroupDao`.
+     * Generates a new UUID, unique identifier for a device. It's used while
+     * extracting a Device from my devices cluster, so I can change the secret
+     * key for this device.
      *
-     * @param application is the context where The Application class in Android
-     * is the base class within an Android app that contains all other
-     * components such as activities and services.
+     * @param oldUuid the old uuid for a specific device, in order to create a
+     * new UUID for it.
      */
-    public DeviceRepository(Application application) {
-        UserDatabase db = UserDatabase.getInstance(application);
-        deviceDao = db.deviceDao();
-        groupDao = db.groupDao();
-        devices = deviceDao.getAll();
-        devicesWithGroups = deviceDao.getAllWithGroups();
-        thisDevice = deviceDao.getThisDevice();
-    }
+    void generateUuid(UUID oldUuid);
 
     /**
-     * {@inheritDoc}
+     * Gets all the devices stored in the database
      *
-     * @param device object to be inserted.
+     * @return list of devices as live data.
      */
-    public void insert(Device device) {
-        UserDatabase.databaseWriteExecutor.execute(() -> deviceDao.insert(device));
-    }
-
-    public void update(Device device) {
-        UserDatabase.databaseWriteExecutor.execute(() -> deviceDao.update(device));
-    }
+    LiveData<List<Device>> getAll();
 
     /**
-     * Deletes a specific device from database. We use it to remove
-     * a device with its groups including whitelisted applications.
+     * Gets all devices with its groups from database using relationship
+     * `DeviceWithGroups`.
      *
-     * @param deviceUuid is the unique identifier for a device to
-     * access it in database.
+     * @return list of devices with its groups as live data.
      */
-    public void delete(UUID deviceUuid) {
-        UserDatabase.databaseWriteExecutor.execute(() -> deviceDao.delete(deviceUuid));
-    }
-
-    public void generateUuid(UUID oldDeviceUuid) {
-        UserDatabase.databaseWriteExecutor.execute(() -> {
-            UUID newDeviceUuid = UUID.randomUUID();
-            groupDao.replaceDeviceUuid(oldDeviceUuid, newDeviceUuid);
-            deviceDao.setUuid(oldDeviceUuid, newDeviceUuid);
-        });
-    }
-
-    public LiveData<List<Device>> getAll() {
-        return devices;
-    }
-
-    public LiveData<List<DeviceWithGroups>> getAllWithGroups() {
-        return devicesWithGroups;
-    }
-
-    public LiveData<Device> getThisDevice() {
-        return thisDevice;
-    }
+    LiveData<List<DeviceWithGroups>> getAllWithGroups();  // TODO maybe deleted
+    LiveData<Device> getThisDevice();
 }
